refactor(core): break races/feats import cycle
Extract HP bonus helpers to hp_bonuses.py and expose public race API
(get_race_and_subrace, grants_as_features, collect_race_features).
Remove dead language branch in feat mechanics grants.
